+++ FE/src/components/game/video/VideoItem.tsx
@@ -4,11 +4,9 @@ import VideoCameraIcon from '@/components/common/icons/VideoCameraIcon';
 import VideoCameraSlashIcon from '@/components/common/icons/VideoCameraSlashIcon';
 import { ROLE, Role } from '@/constants/role';
 import { Situation } from '@/constants/situation';
-import { useAuthStore } from '@/stores/authStore';
-import { useSocketStore } from '@/stores/socketStore';
+import { useVideoItem } from '@/hooks/game/chatting/useVideoItem';
 import { GamePublisher } from '@/types/gamePublisher';
 import { GameSubscriber } from '@/types/gameSubscriber';
-import { useEffect, useRef } from 'react';
 import { FaMicrophone, FaMicrophoneSlash } from 'react-icons/fa';
 
 interface VideoItemProps {
@@ -30,69 +28,14 @@ const VideoItem = ({
   target,
   setTarget,
 }: VideoItemProps) => {
-  const { nickname } = useAuthStore();
-  const { socket } = useSocketStore();
-  const videoRef = useRef<HTMLVideoElement>(null);
-
-  const handleClick = () => {
-    if (!situation || !isPublisherAlive || !gameParticipant.isCandidate) {
-      return;
-    }
-
-    switch (situation) {
-      case 'VOTE':
-        if (target !== null) {
-          socket?.emit('cancel-vote-candidate', {
-            roomId,
-            from: nickname,
-            to: target,
-          });
-        }
-
-        if (target === gameParticipant.nickname) {
-          setTarget(null);
-          return;
-        }
-
-        setTarget(gameParticipant.nickname);
-        socket?.emit('vote-candidate', {
-          roomId,
-          from: nickname,
-          to: gameParticipant.nickname,
-        });
-        break;
-      case 'MAFIA':
-        socket?.emit('select-mafia-target', {
-          roomId,
-          from: nickname,
-          target: gameParticipant.nickname,
-        });
-        break;
-      case 'DOCTOR':
-        setTarget(gameParticipant.nickname);
-        socket?.emit('select-doctor-target', {
-          roomId,
-          from: nickname,
-          target: gameParticipant.nickname,
-        });
-        break;
-      case 'POLICE':
-        socket?.emit('police-investigate', {
-          roomId,
-          police: nickname,
-          criminal: gameParticipant.nickname,
-        });
-        break;
-      default:
-        break;
-    }
-  };
-
-  useEffect(() => {
-    if (videoRef.current && gameParticipant.participant) {
-      gameParticipant.participant.addVideoElement(videoRef.current);
-    }
-  }, [gameParticipant.participant]);
+  const { videoRef, handleClick } = useVideoItem(
+    roomId,
+    isPublisherAlive,
+    gameParticipant,
+    situation,
+    target,
+    setTarget,
+  );
 
   return (
     <div
